// ms_projects/src/models/users/IUser.interface.ts
import { Types } from "mongoose";

export type TUser = {
  _id:Types.ObjectId;
  email: string;
  password: string;
  role: string;
  isSuperUser: boolean;
  token: string;
  active: boolean;
  isBanned: boolean;
  firstLogin: boolean;
  profile: Types.ObjectId;
  thirdPartyProvider: string;
};

export interface IUser extends TUser, Document {}